Factor out common β-reduction code
